+++ docs/webhooks/events-redemption-rollback-failed.md
@@ -2,13 +2,12 @@
 title: Rollback failed
 type: endpoint
 categorySlug: webhooks
-parentDocSlug: redemption
+parentDocSlug: events-redemption
 slug: events-redemption-rollback-failed
 hidden: false
 order: 2
 ---
-# /EVENTS.REDEMPTION.ROLLBACK.FAILED
-## HTTP method: POST
+# EVENT: "EVENTS.REDEMPTION.ROLLBACK.FAILED"<br />HTTP method: POST
 
 [block:html]
 {
