+++ package.json
@@ -12,6 +12,7 @@
     "fix-wrong-responses-names": "ts-node ./scripts/fix-wrong-responses-names.ts",
     "check-schemas-validity": "ts-node ./scripts/check-schemas-validity.ts",
     "prepare-webhooks-documentation": "ts-node ./scripts/prepare-webhooks-documentation.ts",
+    "prepare-webhooks-documentation-2": "ts-node ./scripts/prepare-webhooks-documentation-2.ts",
     "build-production-openapi": "ts-node ./scripts/build-production-openapi.ts",
     "fix-zero-level-schemas": "ts-node ./scripts/fix-zero-level-schemas.ts",
     "generate-endpoints-coverage-docs": "ts-node ./scripts/generate-endpoints-coverage-docs.ts",
