+++ reference/OpenAPIWebhooks.json
@@ -1,17 +1,17 @@
 {
   "openapi": "3.1.0",
   "info": {
-    "title": "Webhook",
+    "title": "Webhooks",
     "version": "1.0.0"
   },
   "tags": [
     {
-      "name": "Redemption",
-      "description": "Redemption"
+      "name": "Events redemption",
+      "description": "Events redemption"
     },
     {
-      "name": "Customer",
-      "description": "Customer"
+      "name": "Events customer",
+      "description": "Events customer"
     }
   ],
   "paths": {
@@ -19,7 +19,7 @@
       "post": {
         "summary": "Failed",
         "tags": [
-          "Redemption"
+          "Events redemption"
         ],
         "responses": {
           "200": {
@@ -40,7 +40,7 @@
       "post": {
         "summary": "Rollback failed",
         "tags": [
-          "Redemption"
+          "Events redemption"
         ],
         "responses": {
           "200": {
@@ -61,7 +61,7 @@
       "post": {
         "summary": "Rollback succeeded",
         "tags": [
-          "Redemption"
+          "Events redemption"
         ],
         "responses": {
           "200": {
@@ -82,7 +82,7 @@
       "post": {
         "summary": "Succeeded",
         "tags": [
-          "Redemption"
+          "Events redemption"
         ],
         "responses": {
           "200": {
@@ -103,7 +103,7 @@
       "post": {
         "summary": "Created",
         "tags": [
-          "Customer"
+          "Events customer"
         ],
         "responses": {
           "200": {
@@ -124,7 +124,7 @@
       "delete": {
         "summary": "Deleted",
         "tags": [
-          "Customer"
+          "Events customer"
         ],
         "responses": {
           "200": {
@@ -145,7 +145,7 @@
       "post": {
         "summary": "Rewarded main",
         "tags": [
-          "Customer"
+          "Events customer"
         ],
         "responses": {
           "200": {
@@ -166,7 +166,7 @@
       "post": {
         "summary": "Rewarded loyalty_points",
         "tags": [
-          "Customer"
+          "Events customer"
         ],
         "responses": {
           "200": {
@@ -187,7 +187,7 @@
       "put": {
         "summary": "Consents revoked",
         "tags": [
-          "Customer"
+          "Events customer"
         ],
         "responses": {
           "200": {
@@ -208,7 +208,7 @@
       "put": {
         "summary": "Consents given",
         "tags": [
-          "Customer"
+          "Events customer"
         ],
         "responses": {
           "200": {
